2019-6-20

+++ src/main/java/com/mmall/param/AclParam.java
@@ -1,5 +1,6 @@
 package com.mmall.param;
 
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -22,7 +23,7 @@ public class AclParam {
     private String name;
 
     @NotNull(message = "必须指定权限模块")
-    private Integer aclModuleId;
+    private Long aclModuleId;
 
     @Length(min = 6,max = 256,message = "权限点名称需要在6-256")
     private String url;
@@ -30,15 +31,15 @@ public class AclParam {
     @NotNull(message = "必须指定权限点的类型")
     @Min(value = 0,message = "权限点类型不合法")
     @Max(value = 2,message = "权限点类型不合法")
-    private Integer type;
+    private Short type;
 
     @NotNull(message = "必须指定权限点的状态")
     @Min(value = 0,message = "权限点类型不合法")
     @Max(value = 1,message = "权限点类型不合法")
-    private Integer status;
+    private Short status;
 
     @NotNull(message = "必须指定权限点的展示顺序")
-    private Integer seq;
+    private Short seq;
 
     @Length(min = 0,max = 256,message = "权限点名称需要在0-256")
     private String remark;
